check whether the directory to delete exists

+++ prepare_data.py
@@ -112,7 +112,8 @@ def prepare_thumbnails():
 def prepare_shards():
     num_shards = 50
 
-    shutil.rmtree("/storage/kaggle/quickdraw/train_simplified_shards")
+    if os.path.isdir("/storage/kaggle/quickdraw/train_simplified_shards"):
+        shutil.rmtree("/storage/kaggle/quickdraw/train_simplified_shards")
     os.makedirs("/storage/kaggle/quickdraw/train_simplified_shards")
 
     categories = read_categories("/storage/kaggle/quickdraw/categories.txt")
